feat(transaction): refactored transaction logic
#8
