--- blockchain/block.go
@@ -1,55 +1,0 @@
-package blockchain
-
-import (
-	"bytes"
-	"encoding/gob"
-	"log"
-	"time"
-)
-
-// Block represents a block in the blockchain
-type Block struct {
-	Timestamp int64
-	// include tx...
-	Data []byte
-	// hash of the previous block header
-	PrevBlockHash []byte
-	Hash          []byte
-	Nonce         int
-}
-
-func NewBlock(data string, prevBlockHash []byte) *Block {
-	block := &Block{time.Now().Unix(), []byte(data), prevBlockHash, []byte{}, 0}
-
-	// use consensus to generate hash and nonce
-	// TODO can be extend more consensus mechanism
-	pow := NewProofOfWork(block)
-	nonce, hash, err := pow.Run()
-	if err != nil {
-		log.Panic(err)
-	}
-	block.Hash = hash[:]
-	block.Nonce = nonce
-	return block
-}
-
-func (b *Block) Serialize() []byte {
-	var result bytes.Buffer
-	encoder := gob.NewEncoder(&result)
-
-	err := encoder.Encode(b)
-	if err != nil {
-		log.Panic(err)
-	}
-	return result.Bytes()
-}
-
-func DeserializeBlock(d []byte) *Block {
-	var block Block
-	decoder := gob.NewDecoder(bytes.NewReader(d))
-	err := decoder.Decode(&block)
-	if err != nil {
-		log.Panic(err)
-	}
-	return &block
-}
