+++ core/blockchain.go
@@ -1,0 +1,120 @@
+package core
+
+import (
+	"fmt"
+	"github.com/boltdb/bolt"
+	"log"
+	"os"
+)
+
+const dbFile = "blockchain_test.db"
+const blocksBucket = "blocks"
+const genesisCoinbaseData = "The Times 03/Jan/2009 Chancellor on brink of second bailout for banks"
+
+type Blockchain struct {
+	//blocks []*Block
+	tip []byte
+	// TODO to support more db
+	Db *bolt.DB
+}
+
+// MineBlock mines a new block with the provided transactions
+func (chain *Blockchain) MineBlock(transactions []*Transaction) {
+	// get last hash from db
+	var lastHash []byte
+	err := chain.Db.View(func(tx *bolt.Tx) error {
+		b := tx.Bucket([]byte(blocksBucket))
+		lastHash = b.Get([]byte("l"))
+		return nil
+	})
+	if err != nil {
+		log.Panic(err)
+	}
+
+	// create new block and store
+	newBlock := NewBlock(transactions, lastHash)
+	err = chain.Db.Update(func(tx *bolt.Tx) error {
+		b := tx.Bucket([]byte(blocksBucket))
+		err = b.Put(newBlock.Hash, newBlock.Serialize())
+		err = b.Put([]byte("l"), newBlock.Hash)
+		chain.tip = newBlock.Hash
+		return nil
+	})
+	if err != nil {
+		log.Panic(err)
+	}
+}
+
+// CreateBlockchain creates a new core DB
+func CreateBlockchain(address string) *Blockchain {
+	if dbExists() {
+		fmt.Println("Blockchain already exists.")
+		os.Exit(1)
+	}
+
+	var tip []byte
+	// Open a DB file.
+	db, err := bolt.Open(dbFile, 0600, nil)
+	if err != nil {
+		log.Panic(err)
+	}
+
+	err = db.Update(func(tx *bolt.Tx) error {
+		coinbaseTx := NewCoinbaseTx(address, genesisCoinbaseData)
+		genesisBlock := NewGenesisBlock(coinbaseTx)
+		b, err := tx.CreateBucket([]byte(blocksBucket))
+		if err != nil {
+			log.Panic(err)
+		}
+		err = b.Put(genesisBlock.Hash, genesisBlock.Serialize())
+		if err != nil {
+			log.Panic(err)
+		}
+		err = b.Put([]byte("l"), genesisBlock.Hash)
+		if err != nil {
+			log.Panic(err)
+		}
+		tip = genesisBlock.Hash
+		return nil
+	})
+	if err != nil {
+		log.Panic(err)
+	}
+	return &Blockchain{tip, db}
+}
+
+func NewBlockChain() *Blockchain {
+	if dbExists() == false {
+		fmt.Println("No existing blockchain found. Create one first.")
+		os.Exit(1)
+	}
+
+	var tip []byte
+	// Open a DB file.
+	db, err := bolt.Open(dbFile, 0600, nil)
+	if err != nil {
+		log.Panic(err)
+	}
+	err = db.Update(func(tx *bolt.Tx) error {
+		b := tx.Bucket([]byte(blocksBucket))
+		tip = b.Get([]byte("l"))
+		return nil
+	})
+	if err != nil {
+		log.Panic(err)
+	}
+	return &Blockchain{tip, db}
+}
+
+// Iterator returns a BlockchainIterator to iterate over the blocks of the core
+func (chain *Blockchain) Iterator() *BlockchainIterator {
+	return &BlockchainIterator{chain.tip, chain.Db}
+}
+
+func dbExists() bool {
+	if _, err := os.Stat(dbFile); os.IsNotExist(err) {
+		return false
+	}
+
+	return true
+}
